--- a/src/amr/CMakeLists.txt
+++ b/src/amr/CMakeLists.txt
@@ -91,4 +91,5 @@ target_link_libraries(${PROJECT_NAME} PUBLIC
   SAMRAI_tbox
   SAMRAI_xfer
   )
+set_property(TARGET ${PROJECT_NAME} APPEND_STRING PROPERTY LINK_FLAGS " ${PHARE_LINK_FLAGS}")