--- a/src/simulator/CMakeLists.txt
+++ b/src/simulator/CMakeLists.txt
@@ -20,4 +20,4 @@ target_link_libraries(${PROJECT_NAME} PUBLIC
   phare_amr # for mpicc
   phare_diagnostic
   )
-
+set_property(TARGET ${PROJECT_NAME} APPEND_STRING PROPERTY LINK_FLAGS " ${PHARE_LINK_FLAGS}")