… src/core/CMakeLists.txt
… src/initializer/CMakeLists.txt