with the following harris run https://gist.github.com/PhilipDeegan/85c186f1442e72dfac04461bbf78d8c9

we reduce the single core serial execution by about 10%

master runs

```
>  python3 tests/functional/harris/harris_2d_2.py
...
mean advance time = 91.78641986846924   91.79sec
total advance time = 91.78641986846924

>  python3 tests/functional/harris/harris_2d_2.py
...
mean advance time = 89.75974607467651   89.76sec
total advance time = 89.75974607467651

>  python3 tests/functional/harris/harris_2d_2.py
...
mean advance time = 90.6504008769989l   90.65sec
total advance time = 90.6504008769989
```

vs

with PGO

```
>  python3 tests/functional/harris/harris_2d_2.py
...
mean advance time = 83.94535183906555   83.95sec
total advance time = 83.94535183906555

>  python3 tests/functional/harris/harris_2d_2.py
...
mean advance time = 81.98065614700317   81.98sec
total advance time = 81.98065614700317

>  python3 tests/functional/harris/harris_2d_2.py
...
mean advance time = 82.83573865890503   82.84sec
total advance time = 82.83573865890503
```

it does require building from source twice, hence the file `res/cmake_pgo_go.sh`

This likely only covers 2d, and executions via python right now.


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
  - Introduced new options for profile guided optimization (PGO) to enhance performance.
  - Added a bash script to streamline CMake configuration and build processes, including performance profiling and Python tests.

- **Enhancements**
  - Improved build configuration by appending link flags to various targets for better optimization and consistency.

- **Bug Fixes**
  - Fixed potential issues with uninitialized `Particle` struct members by setting default values.

- **Documentation**
  - Updated comments in the codebase to include references to related GitHub issues for better context and tracking.

- **Refactor**
  - Modified method signatures to use constant parameters where appropriate, ensuring better code safety and clarity.
  - Streamlined update methods by removing unused parameters, leading to cleaner code and potentially improved performance.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->