@PhilipDeegan 2023-12-10
having tried this on TC and another laptop I don't see any gain there, so it could be CPU specific to some degree